missed files

+++ ThinIce/app/src/main/java/com/udtech/thinice/model/users/User.java
@@ -2,13 +2,12 @@ package com.udtech.thinice.model.users;
 
 import com.orm.SugarRecord;
 
-import java.io.Serializable;
 import java.util.Date;
 
 /**
  * Created by JOkolot on 03.11.2015.
  */
-public class User extends SugarRecord<User>{
+public class User extends SugarRecord<User> {
     private long twitterId;
     private long facebookId;
     private String imageUrl;
@@ -18,8 +17,8 @@ public class User extends SugarRecord<User>{
     private String email;
     private Boolean sex;//true - male, false female;
     private Date dateOfBirth;
-    private Long weight;
-    private Long height;
+    private float weight;
+    private float height;
 
     public User() {
     }
@@ -96,26 +95,26 @@ public class User extends SugarRecord<User>{
         this.dateOfBirth = dateOfBirth;
     }
 
-    public Long getWeight() {
+    public float getWeight() {
         return weight;
     }
 
-    public void setWeight(Long weight) {
+    public void setWeight(float weight) {
         this.weight = weight;
     }
 
-    public Long getHeight() {
+    public float getHeight() {
         return height;
     }
 
-    public void setHeight(Long height) {
+    public void setHeight(float height) {
         this.height = height;
     }
 
     @Override
     public boolean equals(Object o) {
-        if(o instanceof  User)
-            return ((User)o).getId() == getId();
+        if (o instanceof User)
+            return ((User) o).getId() == getId();
         return super.equals(o);
     }
 }
